Split the buffer-build module along the seam it already had
Two passes and both their test suites had grown to three and a half
thousand lines in one file, which is well past the size that asks for a
look. The seam was already written in the file: the recognisers, then
the loop whose accumulator is the answer, then the tests.

So `mod.rs` keeps the joined-builder pass and the recognisers the two
share, `list_build.rs` takes the collecting-loop pass built on them, and
`tests.rs` takes both suites the way every other module in this tree
keeps them. Nothing moved between the passes and nothing changed
meaning: the second one reads its neighbours through `use super::*`
exactly as it read them when they were above it in the same file.
